Add a generic data type converter to the `Cursor` object
This will allow converting fetched data from CrateDB data types to
Python data types in different ways. Its usage is completely optional.
When not used, the feature will not incur any overhead.

There is also a `DefaultTypeConverter`, which aims to become a sane
default choice when looking at this kind of convenience. It will enable
the user to work with native Python data types from the start, for all
database types where this makes sense, without needing to establish the
required set of default value converters on their own.

If the `DefaultTypeConverter` does not fit the user's aims, it is easy
to define custom type converters, possibly reusing specific ones from
the library.