Tests: Prune list of provided `test.globs` for doctests
Remove any kinds of symbol aliases from the doctest launcher established
per `test.globs`, in order not to obfuscate them away from the users
reading the examples, and likely aiming to just copy/paste the presented
code snippets.

This is to remove any kinds of idioms or shortcuts from the
doctest documentation, in order to make it clear what is happening.

Also, it explains where object or response mocks are used, and why.